+++ mefit/Program.cs
@@ -23,12 +23,12 @@ namespace Mac_EFI_Toolkit
     static class Program
     {
         internal static readonly string appBuild = $"{Application.ProductVersion}.230620.1842";
+        internal static readonly string appChannel = "BETA";
         internal static string appDirectory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
         internal static string fsysDirectory = Path.Combine(appDirectory, "fsys_stores");
         internal static string buildsDirectory = Path.Combine(appDirectory, "builds");
         internal static string appName = Assembly.GetExecutingAssembly().Location;
         internal static string draggedFile = string.Empty;
-        internal static readonly string closeChar = "\xE947";
 
         #region Private Members
         private static NativeMethods.LowLevelKeyboardProc _kbProc = HookCallback;
@@ -53,6 +53,7 @@ namespace Mac_EFI_Toolkit
         internal static Font FONT_MDL2_REG_20;
         internal static Font FONT_MDL2_REG_14;
         internal static Font FONT_MDL2_REG_12;
+        internal static Font FONT_MDL2_REG_10;
         internal static Font FONT_MDL2_REG_9;
         #endregion
 
@@ -75,6 +76,7 @@ namespace Mac_EFI_Toolkit
             // Font Data
             byte[] fontData = Properties.Resources.segmdl2;
             FONT_MDL2_REG_9 = new Font(LoadFontFromResource(fontData), 9.0F, FontStyle.Regular);
+            FONT_MDL2_REG_10 = new Font(LoadFontFromResource(fontData), 10.0F, FontStyle.Regular);
             FONT_MDL2_REG_12 = new Font(LoadFontFromResource(fontData), 12.0F, FontStyle.Regular);
             FONT_MDL2_REG_14 = new Font(LoadFontFromResource(fontData), 14.0F, FontStyle.Regular);
             FONT_MDL2_REG_20 = new Font(LoadFontFromResource(fontData), 20.0F, FontStyle.Regular);